# diff: optimize allocation of histogram entries for unique words